feat(template): move customer-account OAuth handling into proxy.ts
Introduce a Next 16 proxy.ts that owns the customer-account OAuth paths
(/account/login|authorize|refresh|logout) via Hydrogen's handleShopifyRoutes
and createCustomerAccountServerHandlers, replacing the dedicated
app/account/[action]/route.ts handler.

- Broad matcher excluding _next assets, favicon, robots, sitemap, .well-known
- Scoped to auth paths (gated by isAuthEnabled) to avoid per-request session
  and storefront client cost; no Hydrogen cart handlers (template keeps its
  server-action cart)
- Forwards request/response headers for non-auth routes
- Delete now-redundant app/account/[action]/route.ts
- Update apps/template/AGENTS.md and docs authentication anatomy to reference
  proxy.ts ownership
