+++ apps/docs/content/docs/anatomy/authentication.mdx
@@ -60,7 +60,7 @@ Use `isCustomerLoggedIn()` for read-only UI state, `requireCustomerSession()` fo
 
 ## Out of the box
 
-- **Auth routes** — `/account/login` (GET, starts PKCE and redirects to Shopify), `/account/authorize` (GET, validates the OAuth callback), `/account/refresh` (GET, rotates an expired access token), and `/account/logout` (POST, clears the session). All four are installed via Hydrogen's standard handlers through `handleShopifyRoutes` in `app/account/[action]/route.ts`.
+- **Auth routes** — `/account/login` (GET, starts PKCE and redirects to Shopify), `/account/authorize` (GET, validates the OAuth callback), `/account/refresh` (GET, rotates an expired access token), and `/account/logout` (POST, clears the session). All four are installed via Hydrogen's standard handlers through `handleShopifyRoutes` in `proxy.ts`, which intercepts the customer-account OAuth paths before routing.
 - **Account pages** — `/account` (redirects to profile), `/account/profile` (edit name, read-only email), `/account/orders` (paginated order history), `/account/orders/[id]` (order detail), and `/account/addresses` (address book with create, edit, delete, and default selection). The `(authenticated)` route group gates these without claiming the auth-handler paths.
 - **Customer Account API data** — a separate GraphQL endpoint and schema from the Storefront API. `lib/shopify/customer-account.ts` wraps Hydrogen's `createCustomerAccountClient`; `lib/shopify/operations/customer.ts` queries and mutates profile, orders, and addresses, resolving a usable token through `requireCustomerAccessToken()` on every call; `lib/shopify/transforms/customer.ts` maps responses to domain types. Order and profile pages are Server Components in Suspense; address and profile editing use client forms backed by server actions in `lib/customer/action.ts` that validate input, surface Shopify `userErrors`, and revalidate affected paths. Customer responses are personalized and never publicly cached.
